+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable match that matches when a value is null (#5)
 - Enable match that matches when a value is not null (#11)
 - Enable match that matches when a value equals a given reference value using an equality comparer (#7)
+- Enable match that matches when a value equals a given reference value using a comparer (#25)
 - Enable match that matches when a value does not equal a given reference value using an equality comparer (#9)
 - Enable match that matches when a value is the same as a given reference value (#13)
 - Enable match that matches when a value is not the same as a given reference value (#15)